Added pagination to notes, comments and replies!

+++ src/Infrastructure/Repositories/CommentRepository.cs
@@ -1,24 +1,19 @@
-﻿using Microsoft.AspNetCore.Http.HttpResults;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
-using PetProject.Application.DTOs.Requests;
-using PetProject.Application.DTOs.Responses;
-using PetProject.Application.Services.Interfaces;
+﻿using Microsoft.EntityFrameworkCore;
 using PetProject.Domain.Entities;
-using PetProject.Infrastructure.EfContext;
+using PetProject.Domain.Repository;
 
-namespace PetProject.Application.Services.Impl;
+namespace PetProject.Infrastructure.Repositories;
 
 public class CommentRepository : ICommentRepository
 {
-    private readonly EfContext _context;
+    private readonly EfContext.EfContext _context;
 
-    public CommentRepository(EfContext context)
+    public CommentRepository(EfContext.EfContext context)
     {
         _context = context;
     }
 
-    public async Task<Comment?> GetCommentById(int id)
+    public async Task<Comment> GetCommentById(int id)
     {
         return await _context.Comment.AsNoTracking()
             .FirstAsync(p => p.CommentId == id);
@@ -28,6 +23,7 @@ public class CommentRepository : ICommentRepository
     {
         return await _context.Comment
             .AsNoTracking()
+            .OrderBy(p => p.PublishedOn)
             .Include(p => p.Replies)
             .FirstOrDefaultAsync(p => p.CommentId == id);
     }
@@ -36,7 +32,7 @@ public class CommentRepository : ICommentRepository
     {
         var note = await _context.Notes.Include(p => p.Comments)
             .FirstOrDefaultAsync(p => p.NoteId == comment.NoteId);
-        note.Comments.Add(comment);
+        note?.Comments.Add(comment);
         await _context.SaveChangesAsync();
         return comment;
     }
@@ -45,7 +41,7 @@ public class CommentRepository : ICommentRepository
     {
         Comment? comment = await _context.Comment.AsNoTracking()
             .FirstOrDefaultAsync(p => p.CommentId == id);
-        _context.Comment.Remove(comment);
+        if (comment != null) _context.Comment.Remove(comment);
         await _context.SaveChangesAsync();
     }
 
@@ -55,5 +51,16 @@ public class CommentRepository : ICommentRepository
             .FirstOrDefaultAsync(p => p.CommentId == commentId);
         comment?.LikeComm();
         await _context.SaveChangesAsync();
+    }
+
+    public async Task<List<Comment>> GetCommentPagination(int pageNumber, int pageSize, int noteId)
+    {
+        return await _context.Comment
+            .AsNoTracking()
+            .OrderBy(p => p.PublishedOn)
+            .Where(p => p.NoteId == noteId)
+            .Skip((pageNumber - 1) * pageSize)
+            .Take(pageSize)
+            .ToListAsync();
     }
 }
